move GenerateInterfaceName to nanoid package
